+++ Desafio_Mestre/Desafio_Mestre.c
@@ -42,8 +42,8 @@ int main() {
     printf("Digite o numero de pontos turisticos: ");
     scanf("%d", &pontos_turisticos_carta1);
 
-    densidade_populacional_carta1 = populacao_carta1 / area_em_km2_carta1;
-    pib_per_capita_carta1 = pib_carta1 / populacao_carta1;
+    densidade_populacional_carta1 = (float)populacao_carta1 / area_em_km2_carta1;
+    pib_per_capita_carta1 = pib_carta1 / (float)populacao_carta1;
     poder_carta1 = (poder_carta1 + area_em_km2_carta1 + pib_carta1 + pontos_turisticos_carta1 + pib_per_capita_carta1) + !densidade_populacional_carta1;
 
     printf(
@@ -104,8 +104,8 @@ int main() {
     printf("Digite o numero de pontos turisticos: ");
     scanf("%d", &pontos_turisticos_carta2);
     
-    densidade_populacional_carta2 = populacao_carta2 / area_em_km2_carta2;
-    pib_per_capita_carta2 = pib_carta2 / populacao_carta2;
+    densidade_populacional_carta2 = (float)populacao_carta2 / area_em_km2_carta2;
+    pib_per_capita_carta2 = pib_carta2 / (float)populacao_carta2;
     poder_carta2 = (poder_carta1 + area_em_km2_carta1 + pib_carta1 + pontos_turisticos_carta1 + pib_per_capita_carta1) + !densidade_populacional_carta1;
 
     printf(
